+ extracted TlsHandler.MediationStream.cs
